# app/static/swagger.json
{
    "info": {
        "title": "My Awesome API",
        "version": "1.0.42",
        "description": "You know, for devs"
    },
    "paths": {
        "/api/users/{id}": {
            "get": {
                "summary": "User endpoint.",
                "description": "Get user by ID.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "ID of the user",
                        "type": "integer",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User object to be returned.",
                        "schema": {
                            "$ref": "#/components/schemas/User"
                        },
                        "examples": {
                            "about_me": {
                                "value": "About me"
                            },
                            "followed_count": {
                                "value": 0
                            },
                            "follower_count": {
                                "value": 0
                            },
                            "id": {
                                "value": 1
                            },
                            "last_seen": {
                                "value": null
                            },
                            "post_count": {
                                "value": 12
                            },
                            "username": {
                                "value": "toto"
                            }
                        }
                    },
                    "404": {
                        "description": "User not found."
                    }
                }
            },
            "put": {
                "summary": "Update an existing user.",
                "description": "Update a existing user.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "required": false,
                        "name": "user",
                        "schema": {
                            "$ref": "#/components/schemas/UpdateUser"
                        },
                        "description": "The user to create."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "404": {
                        "description": "User not found."
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "summary": "User endpoint.",
                "description": "Get users list.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "description": "The page yout want to access",
                        "type": "integer"
                    },
                    {
                        "in": "query",
                        "name": "per_page",
                        "description": "The numbers of User to return",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User object to be returned.",
                        "schema": {
                            "$ref": "#/components/schemas/Users"
                        }
                    },
                    "404": {
                        "description": "User not found."
                    }
                }
            },
            "post": {
                "summary": "Creates a new user.",
                "description": "Creates a new user.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "required": false,
                        "name": "user",
                        "schema": {
                            "$ref": "#/components/schemas/CreateUser"
                        },
                        "description": "The user to create."
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "404": {
                        "description": "User not found."
                    }
                }
            }
        },
        "/api/users/{id}/followers": {
            "get": {
                "summary": "User followers endpoint.",
                "description": "Get followers list.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "description": "The page yout want to access",
                        "type": "integer"
                    },
                    {
                        "in": "query",
                        "name": "per_page",
                        "description": "The numbers of User to return",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User followers object to be returned.",
                        "schema": {
                            "$ref": "#/components/schemas/Users"
                        }
                    },
                    "404": {
                        "description": "User not found."
                    }
                }
            }
        },
        "/api/users/{id}/followed": {
            "get": {
                "summary": "User followed endpoint.",
                "description": "Get followed list.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "description": "The page yout want to access",
                        "type": "integer"
                    },
                    {
                        "in": "query",
                        "name": "per_page",
                        "description": "The numbers of User to return",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User followers object to be returned.",
                        "schema": {
                            "$ref": "#/components/schemas/Users"
                        }
                    },
                    "404": {
                        "description": "User not found."
                    }
                }
            }
        },
        "/api/tokens": {
            "get": {
                "summary": "User followed endpoint.",
                "description": "Get followed list.",
                "security": [
                    {
                        "basicAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token to use for other api route",
                        "schema": {
                            "$ref": "#/components/schemas/Token"
                        }
                    },
                    "404": {
                        "description": "User not found."
                    }
                }
            }
        }
    },
    "tags": [],
    "openapi": "3.0",
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "post_count": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "follower_count": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "last_seen": {
                        "type": "string",
                        "format": "date"
                    },
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "about_me": {
                        "type": "string"
                    },
                    "followed_count": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "_links": {
                        "$ref": "#/components/schemas/Links"
                    },
                    "username": {
                        "type": "string"
                    }
                },
                "required": [
                    "about_me",
                    "username"
                ]
            },
            "Users": {
                "type": "object",
                "properties": {
                    "_links": {
                        "$ref": "#/components/schemas/Links"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    },
                    "_meta": {
                        "$ref": "#/components/schemas/Meta"
                    }
                }
            },
            "Links": {
                "type": "object",
                "properties": {
                    "followed": {
                        "type": "string",
                        "format": "url"
                    },
                    "self": {
                        "type": "string",
                        "format": "url"
                    },
                    "followers": {
                        "type": "string",
                        "format": "url"
                    },
                    "avatar": {
                        "type": "string",
                        "format": "url"
                    }
                }
            },
            "Meta": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "per_page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "total_pages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "total_items": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "CreateUser": {
                "type": "object",
                "properties": {
                    "password": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "about_me": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "password",
                    "username"
                ]
            },
            "UpdateUser": {
                "type": "object",
                "properties": {
                    "password": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "about_me": {
                        "type": "string"
                    }
                }
            },
            "Token": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    }
                }
            }
        },
        "parameters": {},
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    },
    "securityDefinitions": {
        "bearerAuth": {
            "description": "",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}